src/slices/car/controllers/deleteCar/deleteCarController.ts
/* eslint-disable no-unsafe-optional-chaining */
import {
  HttpRequest,
  HttpResponse,
  Validation,
  badRequest,
  ok,
} from "@/application/helpers";
import { Controller } from "@/application/infra/contracts";
import { DeleteCar } from "@/slices/car/useCases";

export class DeleteCarController extends Controller {
  constructor(
    private readonly validation: Validation,
    private readonly deleteCar: DeleteCar
  ) {
    super();
  }
  async execute(httpRequest: HttpRequest<any>): Promise<HttpResponse<any>> {
    const errors = this.validation.validate(httpRequest?.query);
    if (errors?.length > 0) {
      return badRequest(errors);
    }
    const carDeleteed = await this.deleteCar({
      fields: { ...httpRequest?.query, createdById: httpRequest?.userId },
      options: {},
    });
    return ok(carDeleteed);
  }
}
